refactor: move raw data parser tests to match src
