+++ src/student_success_tool/dataio/pdp/raw_data.py
@@ -179,34 +179,3 @@ def _maybe_convert_maybe_validate_data(
 
 def _parse_dttm_values(df: pd.DataFrame, *, col: str, fmt: str) -> pd.Series:
     return pd.to_datetime(df[col], format=fmt)
-
-
-# TODO: remove these functions once corresponding tests have been migrated
-
-
-def _uppercase_string_values(df: pd.DataFrame, *, col: str) -> pd.Series:
-    return df[col].str.upper()
-
-
-def _replace_values_with_null(
-    df: pd.DataFrame, *, col: str, to_replace: str | list[str]
-) -> pd.Series:
-    return df[col].replace(to_replace=to_replace, value=None)
-
-
-def _cast_to_bool_via_int(df: pd.DataFrame, *, col: str) -> pd.Series:
-    return (
-        df[col]
-        .astype("string")
-        .map(
-            {
-                "1": True,
-                "0": False,
-                "True": True,
-                "False": False,
-                "true": True,
-                "false": False,
-            }
-        )
-        .astype("boolean")
-    )
